+++ example/lib/src/feature/purchase/purchase_bloc.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:io';
 import 'package:apphud/apphud.dart';
 import 'package:pedantic/pedantic.dart';
 import 'package:apphud/models/apphud_models/apphud_attribution_provider.dart';
@@ -47,6 +46,7 @@ class PurchaseBloc extends Bloc<PurchaseEvent, PurchaseState>
         paywallShown: _mapPaywallShown,
         paywallClosed: _mapPaywallClosed,
         grantPromotional: _mapGrantPromotional,
+        refreshEntitlements: _mapRefreshEntitlements,
       );
 
   void _fetchSubscriptions() {
@@ -267,6 +267,21 @@ class PurchaseBloc extends Bloc<PurchaseEvent, PurchaseState>
       ),
       onError: (e) => printError(
         'grantPromotional(${event.product.productId})',
+        e,
+      ),
+    ));
+  }
+
+  Stream<PurchaseState> _mapRefreshEntitlements(
+    RefreshEntitlements event,
+  ) async* {
+    unawaited(Apphud.refreshEntitlements().then(
+      (value) => printAsJson(
+        'refreshEntitlements()',
+        'success',
+      ),
+      onError: (e) => printError(
+        'refreshEntitlements()',
         e,
       ),
     ));