+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.2.4
+- [Android] Fixed a bug when ApphudListener’s paywallsDidFullyLoad method wasn’t called in some cases.
+- [Android] Added new method refreshEntitlements().
+
+- Dependencies of Native SDK's were updated to:
+  - [Android] 1.6.2
+  - [iOS] 2.8.2
+
 ## 2.2.3
 - [iOS] Payment swizzle was disabled for observer mode.
 - [iOS] Improved interaction with Apple Search Ads attribution.